finetune_embeddings/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
transformers==4.12.0
numpy==1.19.5
tqdm==4.62.3
matplotlib==3.4.3
pyyaml==5.4.1
scikit-learn==0.24.2
tensorboard==2.7.0
seaborn==0.11.2
umap-learn==0.5.1
scipy==1.7.1
captum==0.5.0
optuna==3.0.0
sqlalchemy==2.0.0
wandb==0.12.0
pandas==1.4.0

# Install spaCy model separately with:
# python -m spacy download en_core_web_sm
